+++ src/dtagent.sql/setup/038_p_update_from_configuration.sql
@@ -40,17 +40,19 @@ declare
     PROCEDURE_TIMEOUT INT;
     DATA_RETENTION_TIME_IN_DAYS INT;
 begin
-    SNOWFLAKE_CREDIT_QUOTA := (select DTAGENT_DB.CONFIG.F_GET_CONFIG_VALUE('core.snowflake_credit_quota', 5));
+    --%OPTION:resource_monitor:
+    SNOWFLAKE_CREDIT_QUOTA := (select DTAGENT_DB.CONFIG.F_GET_CONFIG_VALUE('core.snowflake.resource_monitor.credit_quota', 5));
     if (SNOWFLAKE_CREDIT_QUOTA IS NOT NULL) then
         call DTAGENT_DB.CONFIG.P_UPDATE_RESOURCE_MONITOR(:SNOWFLAKE_CREDIT_QUOTA);
     end if;
+    --%:OPTION:resource_monitor
 
     PROCEDURE_TIMEOUT := (select DTAGENT_DB.CONFIG.F_GET_CONFIG_VALUE('core.procedure_timeout', 3600));
     if (PROCEDURE_TIMEOUT IS NOT NULL) then
         execute immediate 'ALTER WAREHOUSE DTAGENT_WH SET STATEMENT_TIMEOUT_IN_SECONDS = ' || :PROCEDURE_TIMEOUT ||  ';';
     end if;
 
-    DATA_RETENTION_TIME_IN_DAYS := (select DTAGENT_DB.CONFIG.F_GET_CONFIG_VALUE('core.snowflake_data_retention_time_in_days', 1));
+    DATA_RETENTION_TIME_IN_DAYS := (select DTAGENT_DB.CONFIG.F_GET_CONFIG_VALUE('core.snowflake.database.data_retention_time_in_days', 1));
     if (DATA_RETENTION_TIME_IN_DAYS IS NOT NULL) then
         execute immediate 'ALTER DATABASE DTAGENT_DB SET DATA_RETENTION_TIME_IN_DAYS = ' || :DATA_RETENTION_TIME_IN_DAYS || ';';
     end if;