+++ src/dtagent/plugins/event_log.sql/admin/071_p_cleanup_event_log.sql
@@ -24,6 +24,7 @@
 --
 -- APP.P_CLEANUP_EVENT_LOG(INT) will remove old event_log entries, with number of hours to retain
 --
+--%OPTION:dtagent_admin:
 use role DTAGENT_OWNER; use database DTAGENT_DB; use warehouse DTAGENT_WH;
 
 drop procedure if exists DTAGENT_DB.APP.P_CLEANUP_EVENT_LOG(INT);
@@ -66,3 +67,4 @@ grant usage on procedure DTAGENT_DB.APP.P_CLEANUP_EVENT_LOG() to role DTAGENT_AD
 use role DTAGENT_VIEWER;
 call DTAGENT_DB.APP.P_CLEANUP_EVENT_LOG();
 */
+--%:OPTION:dtagent_admin